Avoid to lock the Assembly introspected by PowerShell and add new infos for predicates
