+++ Expressif/Functions/Introspection/FunctionIntrospector.cs
@@ -42,16 +42,6 @@ namespace Expressif.Functions.Introspection
             }
         }
 
-        protected IEnumerable<ParameterInfo> BuildParameters(CtorInfo[] ctorInfos)
-            => ctorInfos.SelectMany(x => x.Parameters)
-                        .Select(y => new ParameterInfo(
-                                    y.Name
-                                    , !ctorInfos.All(c => c.Parameters.Any(p => p.Name == y.Name))
-                                    , ctorInfos.First(c => c.Parameters.Any(p => p.Name == y.Name))
-                                                    .Parameters
-                                                    .Single(p => p.Name == y.Name)
-                                                    .Summary
-                            )
-                        ).Distinct();
+        
     }
 }
