+++ generate-info.ps1
@@ -21,27 +21,29 @@ If(Test-Path -LiteralPath $destinationPath\$destinationFile -PathType leaf) {
 $assemblyPath = "Expressif\bin"
 Set-Location $assemblyPath
 $dllfile = "net6.0\Expressif.dll"
+Set-Location "..\..\"
 
 If ((-not (Test-Path -Path "Release\$dllfile")) -or ("Release\$dllfile".CreationTime -lt "Debug\$dllfile".CreationTime)) {
     $directory = "Debug"    
 } else {
     $directory = "Release"   
 }
-Add-Type -Path "$directory\$dllfile"
-
-Set-Location "..\..\"
-
 Write-Host "Generating JSON for $($class.ToLower())s based on $assemblyPath\$directory\$dllfile"
-$elapsed = Measure-Command -Expression {
-    $TextInfo = (Get-Culture).TextInfo
-    $locator = New-Object -TypeName "Expressif.$($TextInfo.ToTitleCase($class))s.Introspection.$($TextInfo.ToTitleCase($class))Introspector"
-    $functions = $locator.Locate() | Sort-Object ListingPriority | Select-Object -Property Name, Aliases, Scope, Summary, Parameters
-    Write-Host  "`t$($functions.Count) $($class.ToLower()) identified"
-    $functions | ForEach-Object {Write-Host "`t`t$($_.Name)"}
-    $functions | ConvertTo-Json -depth 4 | Out-File "$destinationPath\$destinationFile"
-}
-Write-Host  "File created at $destinationPath\$destinationFile in $($elapsed.TotalSeconds) seconds"
 
+$job = Start-Job -ScriptBlock { param($fullDllPath, $class, $destination)
+    Add-Type -Path "$fullDllPath"
+    $elapsed = Measure-Command -Expression {
+        $TextInfo = (Get-Culture).TextInfo
+        $locator = New-Object -TypeName "Expressif.$($TextInfo.ToTitleCase($class))s.Introspection.$($TextInfo.ToTitleCase($class))Introspector"
+        $functions = $locator.Locate() | Sort-Object ListingPriority | Select-Object -Property Name, Aliases, Scope, Summary, Parameters
+        Write-Host  "`t$($functions.Count) $($class.ToLower()) identified"
+        $functions | ForEach-Object {Write-Host "`t`t$($_.Name)"}
+        $functions | ConvertTo-Json -depth 4 | Out-File "$destination"
+    }
+    Write-Host  "File created at $destination in $($elapsed.TotalSeconds) seconds"
+} -Args "$assemblyPath\$directory\$dllfile", $class, "$destinationPath\$destinationFile"
+Wait-Job $job
+Receive-Job $job
 
 ########### Check if it's useful to report a change #############
